pyproject.toml
[project]
name = "mcp-windows"
version = "0.0.1" 
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "TerminalMan", email = "84923604+SecretiveShell@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.2.0",
    "mcp>=1.6.0",
    "pywin32>=310",
    "winrt-runtime>=3.1.0",
    "winrt-windows-data-xml-dom>=3.1.0",
    "winrt-windows-foundation>=3.1.0",
    "winrt-windows-foundation-collections>=3.1.0",
    "winrt-windows-media-control>=3.1.0",
    "winrt-windows-ui-notifications>=3.1.0",
]

[project.scripts]
mcp-windows = "mcp_windows:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
